requirements.txt
antialiased_cnns==0.3
hydra-core==1.3.2
numpy==1.26.4
omegaconf==2.2.3
open_clip_torch==2.14.0
pandas==2.1.4
Pillow>=10.3.0
pytest==6.2.4
pytorch_lightning==1.6.0
pyaml==21.10.1
scikit_learn==1.4.2
setuptools==65.5.1
submitit==1.3.3
timm==0.8.17.dev0
torchmetrics==0.9.3
torch == 2.2.2+cu121
torchvision==0.17.2+cu121
tqdm==4.65.0
transformers>=4.38.0
wandb==0.12.14
tensorboard==2.6.0
pytorch-cuda==12.1 
#torch>=1.13.1+cu113
#torchvision==0.14.1+cu113
# Test
